requirements.txt
rembg[cpu]
Pillow
streamlit
